feat: 更換profile/info的data結構到account/info內

+++ lib_common/src/main/java/arch/cayenne/lib/common/data/constants/BaseCurrencyData.kt
@@ -17,7 +17,7 @@ sealed class BaseCurrencyData {
         val id: Int,
         val icon: String,
         val currencyName: String,
-        val amount: Long,
+        val amount: Double,
         val amountStr: String,
         val unit: String
     ) : BaseCurrencyData()
